Fix major issue with /adventure
This fix refactors /adventure to fix the bug where most adventure
variables are synced/saved into the actual file of the code, meaning the
adventure data from different instances can and will override each
other.

The fix also splits most adventure logic into a combat handler in hopes
of merging common functionality with /duel (as it needs to be refactored
in a similar way).