Rename AdventureData to AdventureManager
Rename the AdventureData class in CombatHandler to AdventureManager and
rename all adventureData variables/parameters to adventureManager.

Similar to DuelManager, some functions from CombatHandler and the embeds
from EmbedData will be merged into AdventureData later to prevent
making cluttered code from trying to merge and handle adventures and
duels in the CombatHandler functions. Thus, to prepare this change, the
class and its instances are renamed.