Remove dummy ability button in duel
When abilityButtons is initialized in DuelManager, a disabled dummy
button is created which is then displayed at the start of the duel. This
was created as ability buttons and cooldown texts used to not display on
duel start.

As the ability UI now properly updates on duel start, the dummy button
is replaced as soon as updateAbilityUI is called. So, remove the dummy
button as it is no longer needed.