Move abilityCounts to AdventureManager
Functions that handle the player's and opponent's actions locally
initialize abilityCounts when the function is run. Apparently it isn't
defined and saved in AdventureManager.

Since I'm about to extract all ability count handling code to
updateAbilityCounts in AdventureManager later, I'm moving abilityCounts
for both the player and opponent to AdventureManager so it can be
accessed throughout adventure code more easily.