Add updateAbilityCounts to AdventureManager
This function does practically the same thing as updateAbilityCounts in
DuelManager. The only difference is that this function takes in a stand
parameter because, unlike in duels, there is no currentStand as both
player and opponent turns are performed in one interaction. The
parameter is there to identify which stand's abilities are supposed to
be updated.

Also added unit tests for updateAbilityCounts.

Issue: #10, #4