Slightly rewrite updateAbilityUI (duel) unit test
The unit tests define areAbilitiesInCooldown with a value of true. At
the time, I thought it would be fine to assign all abilities in
areAbilitiesInCooldown to true since it is defined the same way during
initialization in DuelManager.

But I think a more "safe" way would be to assign the default values in
the unit test as null. This is because one of the tests check if a
button should be disabled when the cooldown is active. That specific
test checks if areAbilitiesInCooldown returns true for that ability.
Now if, for some reason, the function did not set the value to true when
it should have been, the test would not detect the supposed failed test.
Thus, assigning null as default values can allow the test to flag that
something has gone wrong should areAbilitiesInCooldown return null
values.