Split bot actions to different functions
Separate botTurn into botAttack, botDodge, and botUseAbility for cleaner
and easier unit testing. botTurn now only handles the overarching logic
for when to execute each action.

Since the bot prioritizes using its abilities over using basic attacks,
it uses hasUsedAbility to check if it should still run botAttack. Now
that the attack and ability code are separated, botUseAbility returns a
bool that is assigned to hasUsedAbility.