Fix modifiers applying to self during stopped time
Whenever the bot uses an ability that would supposedly apply a modifier
to the player's defense during stopped time, that modifier is instead
applied to the bot's defense as long as the bot uses that ability before
its last turn of time stop. To fix this, only update the defense
modifier if either time is not stopped or the bot is on its last turn of
time stop.