Replace data param with object deconstruction
The data parameter in setTurnText has an object as its default value.
The object contains default values for damage, abilityText, and
isConfused. However, overriding the default for one property will create
a new object without the defaults for the other properties. To fix this,
replace the data parameter with object deconstruction. This allows each
variable to keep their default values.