Move setTurntext from botAttack to botTurn
To simplify and make it easier to debug and unit test botTimeStopTurn, I
am reducing the responsibilities of the botAttack, botDodge, and
botUseAbility. Thus, move the setTurnText logic for botTurn to handle.
botAttack will now return the necessary variables for botTurn to call
setTurnText.

botDodge and botUseAbility will also be rewritten in a similar way.

Also updates the unit tests for botAttack and botTurn.

Issue: #5, #4