Rewrite orderEmbedDisplay for player time stop
orderEmbedDisplay used to not check if opponentTurnEmbed has content
whenever the opponent goes first because the assumption is that the
opponent would have an action regardless. What I didn't account for back
then was that the player performing a time stop breaks that assumption
as the opponent would not do a turn on that round. Update
orderEmbedDisplay to check if opponentTurnEmbed has content.

Also updates orderEmbedDisplay's unit tests.