Rewrite IsAdventuring check in acceptAdventure
The conditional statement that checks is a player is already in an
adventure tries to get the IsAdventuring property. However, the findOne
function is asynchronous and is called with await. Thus, to prevent any
future crashes, IsAdventuring is accessed after the promise from findOne
is resolved (which is done by wrapping the await and findOne call in
parenthesis and accessing IsAdventuring outside of that grouping).