Fix turn text empty for some player abilities
There has been a bug with some abilities (in this case, Dome of Aegis
was one of such abilities) where it would have a turn embed with no
text, which would then crash the bot. To fix it, add an else block in
the ability type conditional block.

Also adds unit tests for useAbility.

Issue: #28, #23