Fix IsAdventuring set to true before accepting
Apparently, I should've noticed this bug way back then if not for my
wrong code. Ever since I changed how the bot fetched IsAdventuring in
acceptAdventure, the bug started happening.

In short, the bot already sets IsAdventuring to true the moment a player
runs /adventure, meaning it stops the player from accepting the
adventure since the code thinks that the player is already in one.

To fix this, move the code that sets IsAdventuring from adventure.js to
AdventureButtonsPlus.js.

Also updates the unit tests for acceptAdventure accordingly.