Add unit tests for giveRewards (duel)
While trying to fix the reward embed error (Issue #39), I tested if
giveRewards was the problem and didn't set the reward embed's
description. Thus, I made unit tests which surprisingly passed, meaning
the bug lies not in giveRewards but the piece of code calling/handling
it. Regardless, add the unit tests for giveRewards in DuelManager.