Fix bot crash on two-way duel invite
The bot crashes whenever a two-way duel invite happens and one person
tries to accept the latter player's duel invite. What should happen is
the bot edits the embed to a message saying that the player is already
inviting someone. But for some reason, instead of an editReply call,
that code snippet had an update call, which wouldn't work with the
current deferUpdate flow.